refactor(tasks): drop the speculative CacheRef byte source from the queued fetch
`streamSettledBody` carried a second, unreachable byte source: a queued
worker sinking the response body to a shared cache and returning a
`CacheRef` at `body`, which the task re-read and re-emitted as deltas.
Nothing produces that ref. `FetchUrlJob` finishes with
`materializeDerivedPort(...)` plus `metadata`, and that helper returns `{}`
for `response_type: "stream"`, so `body` is never a ref; no queue storage
implements `subscribeToStream`; and `outputStreamResolver` is injected
nowhere. Both of the branch's loud throws were reachable only from tests
that minted a ref by hand, which is also why its design choices — such as
preferring the deterministic cache slot over the private one — could never
be measured against a real worker.

What is live is the channel-less passthrough: a handle with no `onStream`
still settles the job and yields its output as the `finish`. With the ref
half gone that is the whole method, so it folds back into its caller, and
`resolveBodyRefBacking` goes with it.

`IExecuteContext.cacheRegistry` stays — `hasOutputCache` reads it so the
conditional-request guard sees all three ways a run resolves a cache.
`resolveJobOutputStream` stays public API in `@workglow/task-graph`; only
this importer goes.

`FetchTaskRefStream.test.ts` covered the removed branch and was the only
test exercising the passthrough, so that one case moves to
`FetchTaskQueuedStream.test.ts` rather than disappearing with the file.
